+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/KtlintExtension.kt
@@ -1,5 +1,7 @@
 package org.jlleitschuh.gradle.ktlint
 
+typealias ReporterType = org.jlleitschuh.gradle.ktlint.reporter.ReporterType
+
 /**
  * Extension class for configuring the [KtlintPlugin].
  */
@@ -7,7 +9,7 @@ open class KtlintExtension {
     /**
      * The version of ktlint to use.
      */
-    var version = "0.9.2"
+    var version = "0.15.0"
     /**
      * Enable verbose mode.
      */
@@ -21,6 +23,10 @@ open class KtlintExtension {
      */
     var android = false
     /**
+     * Enable console output mode.
+     */
+    var outputToConsole = true
+    /**
      * Whether or not to allow the build to continue if there are warnings;
      * defaults to {@code false}, as for any other static code analysis tool.
      * <p>
@@ -29,11 +35,24 @@ open class KtlintExtension {
     var ignoreFailures = false
 
     /**
+     * Report output formats.
+     *
+     * Available values: plain, plain_group_by_file, checkstyle, json.
+     *
+     * **Note** for Gradle scripts: for now all values should be uppercase due to bug in Gradle.
+     *
+     * Default is empty.
+     */
+    var reporters: Array<ReporterType> = emptyArray()
+
+    /**
      * Report output format.
      *
      * Available values: plain, plain_group_by_file, checkstyle, json.
      *
      * Default is plain.
      */
-    var reporter: ReporterType = ReporterType.PLAIN
+    @Deprecated(message = "Ktlint introduced multi output support since 0.11.1 version",
+            replaceWith = ReplaceWith("reporters = arrayOf()", "reporter"))
+    var reporter: ReporterType? = ReporterType.PLAIN
 }