+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/KtlintPlugin.kt
@@ -19,6 +19,7 @@ import org.gradle.api.tasks.StopExecutionException
 import org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet
 import kotlin.reflect.KClass
 import net.swiftzer.semver.SemVer
+import org.jlleitschuh.gradle.ktlint.reporter.applyReporters
 
 const val VERIFICATION_GROUP = "Verification"
 const val FORMATTING_GROUP = "Formatting"
@@ -43,8 +44,7 @@ open class KtlintPlugin : Plugin<Project> {
         }
     }
 
-    private fun addKtLintTasksToKotlinPlugin(target: Project,
-                                             extension: KtlintExtension) {
+    private fun addKtLintTasksToKotlinPlugin(target: Project, extension: KtlintExtension) {
         target.pluginManager.withPlugin("kotlin") {
             target.afterEvaluate {
                 val ktLintConfig = createConfiguration(target, extension)
@@ -105,10 +105,10 @@ open class KtlintPlugin : Plugin<Project> {
             }
 
     private fun addAdditionalRunArgs(extension: KtlintExtension, runArgs: MutableSet<String>) {
-        // Add the args to enable verbose and debug mode.
         if (extension.verbose) runArgs.add("--verbose")
         if (extension.debug) runArgs.add("--debug")
-        if (extension.android && SemVer.parse(extension.version).compareTo(SemVer(0, 12, 0)) >= 0) {
+        if (extension.android && SemVer.parse(extension.version) >= SemVer(0, 12, 0)) {
+            // Android option is available from ktlint 0.12.0
             runArgs.add("--android")
         }
     }
@@ -161,7 +161,7 @@ open class KtlintPlugin : Plugin<Project> {
             args(runArgs)
         }.apply {
             this.isIgnoreExitValue = extension.ignoreFailures
-            this.applyReporter(target, extension, sourceSetName)
+            this.applyReporters(target, extension, sourceSetName)
         }
     }